requirements.txt
colorama==0.4.4
pandas==1.2.4
numpy==1.20.3
binance-connector==1.3.0 # Assuming you are using the Binance Connector package